--- Bismuth/Assets/Scripts/제갈도원/SummonChanceSheet/SummonChanceTester.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f8042a9270727184fbcf4ff308349946
